Factor out DeviceMetaData
